Update Camera input handling and clean up vector initialization
Changed `processKeyboardInput` to use `double` for `deltaTime` to improve precision, with proper type casting for calculations. Simplified vector initialization in Primary scene for cleaner code readability.

+++ include/core/Camera.h
@@ -12,7 +12,7 @@ public:
 
   // Camera controls
   void processMouseMovement(float xOffset, float yOffset);
-  void processKeyboardInput(CameraMovement direction, float deltaTime);
+  void processKeyboardInput(CameraMovement direction, double deltaTime);
   void processScroll(float yOffset);
   void lookAt(const glm::vec3 &target);
